LayoutTests/svg/animations/svgenum-animation-3-expected.txt

SVG 1.1 dynamic animation tests

Stretched text
Test SVGLengthAdjustType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.lengthAdjust.animVal is SVGTextContentElement.LENGTHADJUST_SPACING
PASS text.lengthAdjust.baseVal is SVGTextContentElement.LENGTHADJUST_SPACING
PASS text.lengthAdjust.animVal is SVGTextContentElement.LENGTHADJUST_SPACING
PASS text.lengthAdjust.baseVal is SVGTextContentElement.LENGTHADJUST_SPACING
PASS text.lengthAdjust.animVal is SVGTextContentElement.LENGTHADJUST_SPACINGANDGLYPHS
PASS text.lengthAdjust.baseVal is SVGTextContentElement.LENGTHADJUST_SPACING
PASS text.lengthAdjust.animVal is SVGTextContentElement.LENGTHADJUST_SPACINGANDGLYPHS
PASS text.lengthAdjust.baseVal is SVGTextContentElement.LENGTHADJUST_SPACING
PASS text.lengthAdjust.animVal is SVGTextContentElement.LENGTHADJUST_SPACINGANDGLYPHS
PASS text.lengthAdjust.baseVal is SVGTextContentElement.LENGTHADJUST_SPACING
PASS successfullyParsed is true

TEST COMPLETE
